Eliminate dependency of service layer on Azul client
