+++ src/azul/azulclient.py
@@ -36,9 +36,6 @@ from azul import (
     cached_property,
     config,
 )
-from azul.deployment import (
-    aws,
-)
 from azul.es import (
     ESClientFactory,
 )
@@ -50,7 +47,6 @@ from azul.http import (
 )
 from azul.indexer import (
     SourceConfig,
-    SourceRef,
     SourceSpec,
 )
 from azul.indexer.index_queue_service import (
@@ -63,7 +59,7 @@ from azul.indexer.index_service import (
     IndexService,
 )
 from azul.indexer.mirror_service import (
-    MirrorAction,
+    MirrorService,
 )
 from azul.plugins import (
     MetadataPlugin,
@@ -71,8 +67,6 @@ from azul.plugins import (
 )
 from azul.queues import (
     Queues,
-    SQSFifoMessage,
-    SQSMessage,
 )
 from azul.types import (
     JSON,
@@ -108,18 +102,9 @@ class AzulClient(SignatureHelper, HasCachedHttpClient):
     def metadata_plugin(self, catalog: CatalogName) -> MetadataPlugin:
         return self.index_service.metadata_plugin(catalog)
 
-    def mirror_source_message(self,
-                              catalog: CatalogName,
-                              source: SourceRef
-                              ) -> SQSFifoMessage:
-        return SQSFifoMessage(
-            body={
-                'action': MirrorAction.mirror_source.to_json(),
-                'catalog': catalog,
-                'source': source.to_json(),
-            },
-            group_id=source.id
-        )
+    @cached_property
+    def mirror_service(self) -> MirrorService:
+        return MirrorService(schema_url_func=None)
 
     def local_reindex(self, catalog: CatalogName, prefix: str) -> int:
         service = self.index_repository_service
@@ -248,18 +233,6 @@ class AzulClient(SignatureHelper, HasCachedHttpClient):
             'No valid sources specified for any catalog')
         return result
 
-    def mirror_queue(self):
-        name = config.mirror_queue.name
-        return aws.sqs_queue(name)
-
-    def queue_mirror_messages(self, messages: Iterable[SQSMessage]) -> int:
-        rate_limit = float(aws.sqs_fifo_rate_limit)
-        if config.is_in_lambda:
-            rate_limit /= config.mirroring_concurrency
-        return self.queues.send_messages(self.mirror_queue(),
-                                         messages,
-                                         rate_limit=rate_limit)
-
     def delete_all_indices(self, catalog: CatalogName):
         self.index_service.delete_indices(catalog)
 
@@ -379,29 +352,6 @@ class AzulClient(SignatureHelper, HasCachedHttpClient):
         queues = self.queues.get_queues([queue_name])
         length, _ = self.queues.get_queue_lengths(queues)
         return length == 0
-
-    def remote_mirror(self,
-                      catalog: CatalogName,
-                      sources: Iterable[tuple[SourceRef, SourceConfig]]
-                      ):
-        mirror_limit = config.catalogs[catalog].mirror_limit
-        if mirror_limit is not None and mirror_limit < 0:
-            log.info('Not mirroring any files in catalog %r because the file '
-                     'size limit is negative', catalog)
-        else:
-
-            def messages():
-                for source, cfg in sources:
-                    if cfg.mirror:
-                        log.info('Mirroring files in source %r from catalog %r',
-                                 str(source.spec), catalog)
-                        yield self.mirror_source_message(catalog, source)
-                    else:
-                        log.info('Not mirroring any files in source %r from catalog %r because '
-                                 'mirroring is explicitly disabled',
-                                 str(source.spec), catalog)
-
-            self.queue_mirror_messages(messages())
 
     def _get_non_empty_fail_queues(self) -> set[str]:
         return {
