+++ test/integration_test.py
@@ -1732,7 +1732,7 @@ class IndexingIntegrationTest(IntegrationTestCase):
             _delete()
             for _ in range(2):
                 for catalog, sources in sources_by_catalog.items():
-                    self.azul_client.remote_mirror(catalog, sources)
+                    self.azul_client.mirror_service.remote_mirror(catalog, sources)
                 self.azul_client.wait_for_mirroring()
                 self._assert_queues_empty([config.mirror_queue.to_fail.name])
             _delete()
